INTERLOK-699 Added initial.branch to URL specification INTERLOK-699 #time 3h
Added support for a custom URL along the lines of
git@bitbucket.org:adaptris/hgca-interlok?initial.branch=staging
Which means we change the initial branch to "staging" when we do
the initial clone.

This means that we don't need vcs.revision=staging on a clean checkout.

- Things that are not clear
-- If the repo exists when you start up; you end up pulling the "master"
branch due to update()
-- How can you switch branches just by changing the URL (because the
local _copy_DoNotEdit) clone has all the branches
-- by specifying vcs.revision=staging the problem goes away, but it
is still a bit shit to have to specify staging in 2x places.