INTERLOK-700 vcs.revision can now specify a branch name. INTERLOK-700 #time 5h INTERLOK-699 Removed ?initial.branch= support as 700 supersedes 699
If the vcs.revision is a remote branch from origin; then create a
local branch that tracks the remote branch. Then we do a checkout of
that branch (and it works).

Modified GitVCS so that the checkout() method also does an update
afterwards; this means that from a git perspective these operations
happen; so if vcs.revision=staging then what effectively happens is

* git clone gitUrl // checkout master by default
* git branch -b staging -t refs/remotes/origin/staging
* git checkout staging

Updated to latest jgit release.